pyproject.toml
[project]
name = "deepsecurity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "DanielPantoja08", email = "danny0897pant@gmail.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "deepface>=0.0.98",
    "matplotlib>=3.7.0",
    "numpy>=1.23.5,<2.0.0",
    "opencv-python>=4.8.0.74,<4.9.0.0",
    "pandas>=2.0.0",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.46",
    "tensorflow==2.15.0",
    "tf-keras==2.15.0",
    "mtcnn>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ipykernel>=7.2.0",
    "sqlmodel>=0.0.14",
]
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "httpx>=0.27.0",
]

